backend/requirements.txt
Django>=4.2
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=4.2
requests>=2.31
openai>=0.28.0
